requirements.txt
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.5
requests>=2.28.2
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0 